Refactor namespaces Yes this is in the wrong branch. No I don't want to talk about it.
