+++ public/Jailbreak.Formatting/Objects/StringFormatObject.cs
@@ -1,6 +1,8 @@
 ﻿using CounterStrikeSharp.API.Modules.Utils;
 
-namespace Jailbreak.Formatting.Formatting;
+using Jailbreak.Formatting.Core;
+
+namespace Jailbreak.Formatting.Objects;
 
 public class StringFormatObject : FormatObject
 {
