+++ src/Jailbreak.Generic/PlayerState/Behaviors/AliveStateTracker.cs
@@ -1,11 +1,9 @@
-﻿using System.Reflection;
-
-using CounterStrikeSharp.API.Core;
+﻿using CounterStrikeSharp.API.Core;
 using CounterStrikeSharp.API.Core.Attributes.Registration;
 
 using Jailbreak.Public.Behaviors;
 
-namespace Jailbreak.Generic.Behaviors;
+namespace Jailbreak.Generic.PlayerState.Behaviors;
 
 public class AliveStateTracker : BaseStateTracker, IPluginBehavior
 {
