+++ src/Jailbreak.Generic/PlayerState/Behaviors/RoundStateTracker.cs
@@ -1,13 +1,11 @@
-﻿using System.Reflection;
-
-using CounterStrikeSharp.API.Core;
+﻿using CounterStrikeSharp.API.Core;
 using CounterStrikeSharp.API.Core.Attributes.Registration;
 
 using Jailbreak.Public.Behaviors;
 
 using Serilog;
 
-namespace Jailbreak.Generic.Behaviors;
+namespace Jailbreak.Generic.PlayerState.Behaviors;
 
 public class RoundStateTracker : BaseStateTracker, IPluginBehavior
 {
